feat: add THI departments screen and integrate campus life events components
- Introduced a new screen for THI departments with routing and feature flag checks.
- Refactored existing event screens to utilize a new `CampusLifeEventsScreen` component for better code reuse.
- Updated the `CampusLifeOrganizersList` component to handle different sections and pages.
- Enhanced localization for missing descriptions and footer information for departments.
- Removed redundant code and improved overall structure for better maintainability.
