+++ src/app/(screens)/cl-events.tsx
@@ -1,97 +1,13 @@
-import { useQuery } from '@tanstack/react-query'
-import { router, useFocusEffect, useLocalSearchParams } from 'expo-router'
 import type React from 'react'
-import { useCallback, useEffect } from 'react'
-import { InteractionManager, View } from 'react-native'
-import { createStyleSheet, useStyles } from 'react-native-unistyles'
-import ClEventsPage from '@/components/Events/cl-events-page'
-import LoadingIndicator from '@/components/Universal/loading-indicator'
-import { useTransparentHeaderPadding } from '@/hooks/useTransparentHeader'
+import CampusLifeEventsScreen from '@/components/Events/campus-life-events-screen'
 import { CAMPUS_LIFE_PUBLIC_ORGANIZER_KIND_STUDENT_ASSOCIATION } from '@/types/campus-life'
-import { loadCampusLifeEvents, QUERY_KEYS } from '@/utils/events-utils'
-import { pausedToast } from '@/utils/ui-utils'
 
 export default function ClEventsScreen(): React.JSX.Element {
-	const { styles, theme } = useStyles(stylesheet)
-	const headerPadding = useTransparentHeaderPadding()
-	const { tab, openEvent, id } = useLocalSearchParams<{
-		tab?: string
-		openEvent?: string
-		id?: string
-	}>()
-
-	const clEventsResult = useQuery({
-		queryKey: [
-			QUERY_KEYS.CAMPUS_LIFE_EVENTS,
-			CAMPUS_LIFE_PUBLIC_ORGANIZER_KIND_STUDENT_ASSOCIATION
-		],
-		queryFn: () =>
-			loadCampusLifeEvents({
-				organizerKind: CAMPUS_LIFE_PUBLIC_ORGANIZER_KIND_STUDENT_ASSOCIATION
-			}),
-		staleTime: 1000 * 60 * 60, // 60 minutes
-		gcTime: 1000 * 60 * 60 * 24 // 24 hours
-	})
-
-	useEffect(() => {
-		if (tab !== 'sports') {
-			return
-		}
-		router.replace({
-			pathname: '/sports',
-			params: {
-				...(openEvent != null ? { openEvent } : {}),
-				...(id != null ? { id } : {})
-			}
-		})
-	}, [tab, openEvent, id])
-
-	useEffect(() => {
-		if (clEventsResult.isPaused && clEventsResult.data != null) {
-			pausedToast()
-		}
-	}, [clEventsResult.isPaused, clEventsResult.data])
-
-	useFocusEffect(
-		useCallback(() => {
-			if (openEvent === 'true' && id) {
-				InteractionManager.runAfterInteractions(() => {
-					router.setParams({ openEvent: 'false' })
-					router.navigate({
-						pathname: '/events/cl/[id]',
-						params: { id }
-					})
-				})
-			}
-		}, [openEvent, id])
-	)
-
 	return (
-		<View
-			style={[styles.page, { paddingTop: headerPadding + theme.margins.page }]}
-		>
-			{tab === 'sports' ? (
-				<View style={styles.loaderWrapper}>
-					<LoadingIndicator />
-				</View>
-			) : (
-				<ClEventsPage
-					clEventsResult={clEventsResult}
-					organizerKind={CAMPUS_LIFE_PUBLIC_ORGANIZER_KIND_STUDENT_ASSOCIATION}
-					clubsListRoute="/cl-clubs"
-				/>
-			)}
-		</View>
+		<CampusLifeEventsScreen
+			organizerKind={CAMPUS_LIFE_PUBLIC_ORGANIZER_KIND_STUDENT_ASSOCIATION}
+			clubsListRoute="/cl-clubs"
+			enableSportsTabRedirect
+		/>
 	)
 }
-
-const stylesheet = createStyleSheet(() => ({
-	page: {
-		flex: 1
-	},
-	loaderWrapper: {
-		flex: 1,
-		justifyContent: 'center',
-		alignItems: 'center'
-	}
-}))
