+++ src/app/(screens)/thi-departments.tsx
@@ -1,18 +1,22 @@
+import { Redirect } from 'expo-router'
 import type React from 'react'
-import CampusLifeEventsScreen from '@/components/Events/campus-life-events-screen'
+import CampusLifeOrganizersList from '@/components/Events/campus-life-organizers-list'
 import { useIsFeatureEnabled } from '@/hooks'
 import { FeatureFlagKeys } from '@/lib/feature-flags'
 import { CAMPUS_LIFE_PUBLIC_ORGANIZER_KIND_THI_DEPARTMENT } from '@/types/campus-life'
 
-export default function ThiEventsScreen(): React.JSX.Element {
+export default function ThiDepartmentsScreen(): React.JSX.Element {
 	const thiEventsVisible = useIsFeatureEnabled(FeatureFlagKeys.thiEventsVisible)
 
+	if (!thiEventsVisible) {
+		return <Redirect href="/(tabs)" />
+	}
+
 	return (
-		<CampusLifeEventsScreen
+		<CampusLifeOrganizersList
 			organizerKind={CAMPUS_LIFE_PUBLIC_ORGANIZER_KIND_THI_DEPARTMENT}
-			clubsListRoute="/thi-departments"
-			queryEnabled={thiEventsVisible}
-			redirectWhenDisabled="/(tabs)"
+			page="thiEvents"
+			section="departments"
 		/>
 	)
 }
